Allow increasing dimension when within min size
The resizer had a minimum size of 50 pixels for both dimensions,
preventing resizing below this limit. This default limit is removed, and
made configurable. The only limit now is that one can't resize to a
negative size in either dimension.

But if Croppie is initialized with an image that is already smaller than the
configured minimum size in a given dimension it was then impossible to
resize it to a larger size in that dimension.

This also allows increasing the size of a dimension even if the new
dimension size will still be smaller than the minimum size for that
dimension. This at least let's the user get the image up to minimum
size, but it's really left to the consumer of Croppie to ensure an image
initially is larger then the minimum size configured.